requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
appscript==1.2.0
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.4.0
blinker==1.4
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.4
colorlog==6.6.0
cryptography==3.2.1
DingtalkChatbot==1.5.3
execnet==1.9.0
Faker==9.8.3
Flask==1.1.4
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.8.1
lxml==4.9.1
MarkupSafe==2.0.1
mitmproxy==5.3.0
msgpack==1.0.4
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==3.13.0
psutil==5.9.1
psycopg2-binary==2.9.3
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pyDes==2.0.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pytest==7.1.2
pytest-forked==1.3.0
pytest-xdist==2.4.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sortedcontainers==2.2.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.10
urwid==2.1.2
Werkzeug==1.0.1
wsproto==0.15.0
xlrd==2.0.1
xlutils==2.0.0
xlwings==0.27.7
xlwt==1.3.0
zipp==3.8.1
zstandard==0.14.1
